requirements.txt
# VESTIGIUM - Zero Budget Aquatic Biomass Radar
# Requisitos de dependencias

# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
plotly>=5.0.0

# JAX & GPU Computing
jax>=0.3.0
jaxlib>=0.3.0
flax>=0.4.0

# Backend & API
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0

# Data Processing
pandas>=1.3.0
scikit-learn>=0.24.0

# Utilities
pyyaml>=5.4.0
python-dotenv>=0.19.0
pydantic>=1.8.0

# Logging & Monitoring
loguru>=0.6.0
tqdm>=4.60.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.18.0
pytest-cov>=2.12.0

# Development
black>=21.6b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Optional: GPU acceleration
# torch>=1.9.0  # Si usas PyTorch además de JAX
# tensorflow>=2.6.0  # Si necesitas TensorFlow

# Optional: Real-time performance
# numba>=0.54.0  # Para aceleración numérica
